Move null comparison extractors into Nullables
