+++ compiler/src/dotty/tools/dotc/ast/TreeInfo.scala
@@ -656,42 +656,6 @@ trait TypedTreeInfo extends TreeInfo[Type] { self: Trees.Instance[Type] =>
     acc(Nil, tree)
   }
 
-  /** An extractor for null comparisons */
-  object CompareNull with
-
-    /** Matches one of
-     *
-     *    tree == null, tree eq null, null == tree, null eq tree
-     *    tree != null, tree ne null, null != tree, null ne tree
-     *
-     *  The second boolean result is true for equality tests, false for inequality tests
-     */
-    def unapply(tree: Tree)(given Context): Option[(Tree, Boolean)] = tree match
-      case Apply(Select(l, _), Literal(Constant(null)) :: Nil) =>
-        testSym(tree.symbol, l)
-      case Apply(Select(Literal(Constant(null)), _), r :: Nil) =>
-        testSym(tree.symbol, r)
-      case _ =>
-        None
-
-    private def testSym(sym: Symbol, operand: Tree)(given Context) =
-      if sym == defn.Any_== || sym == defn.Object_eq then Some((operand, true))
-      else if sym == defn.Any_!= || sym == defn.Object_ne then Some((operand, false))
-      else None
-
-  end CompareNull
-
-  /** An extractor for comparisons between a path and null. */
-  object ComparePathNull with
-    def unapply(tree: Tree)(given Context): Option[(TermRef, Boolean)] =
-      CompareNull.unapply(tree) match
-        case some @ Some((x, testEqual)) =>
-          x.tpe match
-            case ref: TermRef if ref.isStable => Some((ref, testEqual))
-            case _ => None
-        case none => None
-  end ComparePathNull
-
   /** Is this pattern node a catch-all or type-test pattern? */
   def isCatchCase(cdef: CaseDef)(implicit ctx: Context): Boolean = cdef match {
     case CaseDef(Typed(Ident(nme.WILDCARD), tpt), EmptyTree, _) =>
